Add useAssignment hook
