+++ src/containers/Review/assignment/AssignmentSubmit.tsx
@@ -1,15 +1,8 @@
 import React from 'react'
 import { Button } from 'semantic-ui-react'
 import { useLanguageProvider } from '../../../contexts/Localisation'
-import { useReviewStructureState } from '../../../contexts/ReviewStructuresState'
-import useUpdateReviewAssignment from '../../../utils/hooks/useUpdateReviewAssignment'
-import useReassignReviewAssignment from '../../../utils/hooks/useReassignReviewAssignment'
 import { AssignmentDetails, FullStructure, SectionAssignee } from '../../../utils/types'
-import {
-  ReviewAssignmentStatus,
-  useUpdateReviewAssignmentMutation,
-} from '../../../utils/generated/graphql'
-import { FetchResult } from '@apollo/client'
+import useUpdateAssignment from '../../../utils/hooks/useUpdateAssignment'
 
 interface AssignmentSubmitProps {
   fullStructure: FullStructure
@@ -25,69 +18,11 @@ const AssignmentSubmit: React.FC<AssignmentSubmitProps> = ({
   enableSubmit,
 }) => {
   const { strings } = useLanguageProvider()
-  const { reviewStructuresState } = useReviewStructureState()
-  const { assignSectionsToUser } = useUpdateReviewAssignment(fullStructure)
-  const { reassignSections } = useReassignReviewAssignment(fullStructure)
-  const [updateReviewAssignment] = useUpdateReviewAssignmentMutation()
-
-  const isChanging = (reviewerId: number, assignedSections: SectionAssignee) => {
-    return !!Object.values(assignedSections).find(
-      (section) => section.newAssignee === reviewerId || section.previousAssignee === reviewerId
-    )
-  }
-
-  interface AssignmentPatch {
-    status: ReviewAssignmentStatus
-    assignedSections?: string[]
-  }
-
-  const createAssignmentPatch = (
-    assignment: AssignmentDetails,
-    assignedSections: SectionAssignee
-  ) => {
-    const reviewerId = assignment.reviewer.id
-    const assignedSectionsCodes = new Set(assignment.assignedSections)
-    const removedSections = Object.entries(assignedSections)
-      .filter(([_, { previousAssignee }]) => previousAssignee === reviewerId)
-      .map((section) => section[0])
-    const addedSections = Object.entries(assignedSections)
-      .filter(([_, { newAssignee }]) => newAssignee === reviewerId)
-      .map((section) => section[0])
-
-    removedSections.forEach((code) => assignedSectionsCodes.delete(code))
-    addedSections.forEach((code) => assignedSectionsCodes.add(code))
-
-    const patch: AssignmentPatch = {
-      status:
-        assignedSectionsCodes.size === 0
-          ? ReviewAssignmentStatus.Available
-          : ReviewAssignmentStatus.Assigned,
-    }
-    if (assignedSectionsCodes.size > 0) patch.assignedSections = Array.from(assignedSectionsCodes)
-    return patch
-  }
-
-  const submitAssignments = async () => {
-    const results: Promise<any>[] = []
-    // Deduce which review assignments will change and create patches
-    assignmentsFiltered.forEach((assignment) => {
-      if (isChanging(assignment.reviewer.id, assignedSections)) {
-        const assignmentPatch = createAssignmentPatch(assignment, assignedSections)
-        results.push(
-          updateReviewAssignment({
-            variables: {
-              assignmentId: assignment.id,
-              assignmentPatch,
-            },
-          })
-        )
-      }
-    })
-    await Promise.all(results)
-    fullStructure.reload()
-
-    if (results.some((result) => result?.errors)) throw new Error('Error updating assignments')
-  }
+  const { submitAssignments } = useUpdateAssignment({
+    fullStructure,
+    assignedSections,
+    assignmentsFiltered,
+  })
 
   return (
     <div style={{ marginTop: 10 }}>
